Update 3-netbox.md
When NAPALM is specified as a requirement, pip 21.0.1 fails to satisfy the cryptography package's requirement for setuptools-rust and therefore must be explicitly stated. While it's arguably not the place the NetBox documentation to spell-out requirements for a third party library, the inclusion of NAPALM in requirements will cause the execution of upgrade.sh to fail. I believe this information would be beneficial to those in the community following the documentation step-by-step.